ENG-0000 - Cleanup
- Removed the search API client and related types, including the old query
parser.  Goodbye!
- Removed some unused utility classes.
- Removed related tests and stuff.

Reduces the output package size by about 40%.
