+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@al/core",
-  "version": "1.0.173",
+  "version": "1.0.174",
   "description": "Node Enterprise Packages for Alert Logic (NEPAL) Core Library",
   "main": "./dist/index.cjs.js",
   "types": "./dist/index.d.ts",
