+++ src/common/utility/index.ts
@@ -1,6 +1,5 @@
 export * from "./al-cabinet";
 export * from "./al-globalizer";
-export * from "./al-query-evaluator.types";
 export * from "./al-stopwatch";
 export * from "./al-trigger.types";
 export * from './al-validation.types';
