+++ src/index.ts
@@ -3,7 +3,6 @@ export * from "./configuration/index";
 export * from "./client/index";
 export * from "./session/index";
 export * from "./aims-client/index";
-export * from "./search-client/index";
 export * from "./subscriptions-client/index";
 export * from "./types/index";
 export * from "./error-handler/index";
