--- src/search-client/index.ts
@@ -1,7 +1,0 @@
-export * from "./parser/index";
-export * from "./search-client";
-export { searchClient as AlSearchClient } from "./search-client";
-export * from './search-client.v2';
-export { alSearchClientV2 as AlSearchClientV2 } from './search-client.v2';
-export * from './search-stylist-client';
-export { alSearchStylist as AlSearchStylist } from './search-stylist-client';
