--- src/search-client/parser/index.ts
@@ -1,7 +1,0 @@
-export * from "./clause.types";
-export * from "./common.types";
-export * from "./cursor.types";
-export * from "./operator.types";
-export * from "./parser.types";
-export * from "./query-builder.types";
-export * from "./query.types";
